feat(clothing)/refactor code in clothing
